segmentation_Mask_RCNN/requirements.txt
numpy==1.17.0
keras==2.0.8
scikit-image==0.15.0
imgaug==0.2.9
utils==0.9.0
opencv-python==4.1.0.25
h5py==2.7.0
scipy==0.19.1
Pillow==6.1.0
matplotlib==3.1.1
pandas==0.25.1
pytz==2019.2
